requirements.txt
Flask==2.2.2
Flask_Limiter==3.3.1
Flask_WTF==1.0.1
mysql_connector_repackaged==0.3.1
python_bcrypt==0.3.2
stripe==5.0.0
WTForms==3.0.1
